Display and screenshot changes
I restructured the code for displaying information, so that the logo and system information are stored separately. This allowed for easily displaying multiple drives.

I also changed the code for getting the resolution, and the bounds on screenshots, so that screen shots on systems with multiple monitors where one is offset work correctly.